requirements.txt
discord.py
python-dotenv
assemblyai
yt-dlp
PyNaCl
deepgram-sdk>=3,<5